requirements.txt
google-generativeai
python-dotenv
json-repair
groq